career_backend_openrouter/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.32.3
